Refactor : Career, Tag 엔티티에 memberId 추가, Controller, Service, Repository 수정 & 통합테스트 코드 수정

+++ src/main/java/umc/kkijuk/server/tag/service/TagServiceImpl.java
@@ -4,17 +4,16 @@ package umc.kkijuk.server.tag.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import umc.kkijuk.server.common.domian.exception.OwnerMismatchException;
 import umc.kkijuk.server.common.domian.exception.InvalidTagNameException;
 import umc.kkijuk.server.common.domian.exception.ResourceNotFoundException;
-import umc.kkijuk.server.common.domian.response.ErrorResponse;
+import umc.kkijuk.server.member.domain.Member;
 import umc.kkijuk.server.tag.domain.Tag;
 import umc.kkijuk.server.tag.dto.TagRequestDto;
 import umc.kkijuk.server.tag.dto.TagResponseDto;
 import umc.kkijuk.server.tag.dto.converter.TagConverter;
 import umc.kkijuk.server.tag.repository.TagRepository;
 
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
@@ -23,22 +22,25 @@ public class TagServiceImpl implements TagService {
     private final TagRepository tagRepository;
     @Override
     @Transactional
-    public Tag createTag(TagRequestDto.CreateTagDto request) {
+    public Tag createTag(Member requestMember, TagRequestDto.CreateTagDto request) {
         String tagName = request.getTagName();
         if (tagName==null || tagName.trim().isEmpty() || tagRepository.existsByName(tagName)) {
             throw new InvalidTagNameException("태그 이름은 비워둘 수 없으며, 이미 존재하는 이름은 사용할 수 없습니다.");
         }
-        Tag hashTag = TagConverter.toTag(request);
-        return tagRepository.save(hashTag);
+        Tag newTag = TagConverter.toTag(requestMember.getId(), request);
+        return tagRepository.save(newTag);
     }
     @Override
-    public TagResponseDto.ResultTagDtoList  findAllTags() {
-        return TagConverter.toResultTagDtoList(tagRepository.findAll());
+    public TagResponseDto.ResultTagDtoList  findAllTags(Member requestMember) {
+        return TagConverter.toResultTagDtoList(tagRepository.findAllTagByMemberId(requestMember.getId()));
     }
     @Override
     @Transactional
-    public void delete(Long tagId) {
+    public void delete(Member requestMember, Long tagId) {
         Tag deleteTag = tagRepository.findById(tagId).orElseThrow(() -> new ResourceNotFoundException("Tag", tagId));
+        if(!deleteTag.getMemberId().equals(requestMember.getId())){
+            throw new OwnerMismatchException();
+        }
         tagRepository.delete(deleteTag);
     }
